feat(#640): add one more puzzle
